app/requirements.txt
flask
flask_sqlalchemy
flask_limiter
pyjwt
python-dotenv
dnspython
gunicorn
Werkzeug
requests
